+++ packages/iov-cosmos-sdk/types/index.d.ts
@@ -1,0 +1,4 @@
+export { CosmosCodec } from "./cosmoscodec";
+export { CosmosConnection, TokenConfiguration } from "./cosmosconnection";
+export { createCosmosConnector } from "./cosmosconnector";
+export { BankToken } from "./types";